Отрисовка фильтров и задач через классы
